+++ cli/src/commands/split.rs
@@ -11,16 +11,26 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+use std::collections::HashMap;
 use std::io::Write as _;
 
 use clap_complete::ArgValueCompleter;
+use jj_lib::backend::CommitId;
 use jj_lib::commit::Commit;
 use jj_lib::matchers::Matcher;
 use jj_lib::object_id::ObjectId as _;
 use jj_lib::repo::Repo as _;
+use jj_lib::rewrite::move_commits;
 use jj_lib::rewrite::CommitWithSelection;
+use jj_lib::rewrite::EmptyBehaviour;
+use jj_lib::rewrite::MoveCommitsLocation;
+use jj_lib::rewrite::MoveCommitsTarget;
+use jj_lib::rewrite::RebaseOptions;
+use jj_lib::rewrite::RebasedCommit;
+use jj_lib::rewrite::RewriteRefsOptions;
 use tracing::instrument;
 
+use crate::cli_util::compute_commit_location;
 use crate::cli_util::CommandHelper;
 use crate::cli_util::DiffSelector;
 use crate::cli_util::RevisionArg;
@@ -70,6 +80,40 @@ pub(crate) struct SplitArgs {
         add = ArgValueCompleter::new(complete::revset_expression_mutable),
     )]
     revision: RevisionArg,
+    /// The revision(s) to rebase onto (can be repeated to create a merge
+    /// commit)
+    #[arg(
+        long,
+        short,
+        conflicts_with = "parallel",
+        value_name = "REVSETS",
+        add = ArgValueCompleter::new(complete::revset_expression_all),
+    )]
+    destination: Option<Vec<RevisionArg>>,
+    /// The revision(s) to insert after (can be repeated to create a merge
+    /// commit)
+    #[arg(
+        long,
+        short = 'A',
+        visible_alias = "after",
+        conflicts_with = "destination",
+        conflicts_with = "parallel",
+        value_name = "REVSETS",
+        add = ArgValueCompleter::new(complete::revset_expression_all),
+    )]
+    insert_after: Option<Vec<RevisionArg>>,
+    /// The revision(s) to insert before (can be repeated to create a merge
+    /// commit)
+    #[arg(
+        long,
+        short = 'B',
+        visible_alias = "before",
+        conflicts_with = "destination",
+        conflicts_with = "parallel",
+        value_name = "REVSETS",
+        add = ArgValueCompleter::new(complete::revset_expression_mutable),
+    )]
+    insert_before: Option<Vec<RevisionArg>>,
     /// The change description to use (don't open editor)
     ///
     /// The description is used for the commit with the selected changes. The
@@ -116,11 +160,29 @@ impl SplitArgs {
             self.tool.as_deref(),
             self.interactive || self.paths.is_empty(),
         )?;
+        let use_move_flags = self.destination.is_some()
+            || self.insert_after.is_some()
+            || self.insert_before.is_some();
+        let (new_parent_ids, new_child_ids) = if use_move_flags {
+            compute_commit_location(
+                ui,
+                workspace_command,
+                self.destination.as_deref(),
+                self.insert_after.as_deref(),
+                self.insert_before.as_deref(),
+                "split-out commit",
+            )?
+        } else {
+            Default::default()
+        };
         Ok(ResolvedSplitArgs {
             target_commit,
             matcher,
             diff_selector,
             parallel: self.parallel,
+            use_move_flags,
+            new_parent_ids,
+            new_child_ids,
         })
     }
 }
@@ -130,6 +192,9 @@ struct ResolvedSplitArgs {
     matcher: Box<dyn Matcher>,
     diff_selector: DiffSelector,
     parallel: bool,
+    use_move_flags: bool,
+    new_parent_ids: Vec<CommitId>,
+    new_child_ids: Vec<CommitId>,
 }
 
 #[instrument(skip_all)]
@@ -144,6 +209,9 @@ pub(crate) fn cmd_split(
         matcher,
         diff_selector,
         parallel,
+        use_move_flags,
+        new_parent_ids,
+        new_child_ids,
     } = args.resolve(ui, &workspace_command)?;
     let text_editor = workspace_command.text_editor()?;
     let mut tx = workspace_command.start_transaction();
@@ -155,6 +223,12 @@ pub(crate) fn cmd_split(
     let first_commit = {
         let mut commit_builder = tx.repo_mut().rewrite_commit(&target.commit).detach();
         commit_builder.set_tree_id(target.selected_tree.id());
+        if use_move_flags {
+            commit_builder
+                // Generate a new change id so that the commit being split doesn't
+                // become divergent.
+                .generate_new_change_id();
+        }
         let description = if !args.message_paragraphs.is_empty() {
             let description = join_message_paragraphs(&args.message_paragraphs);
             if !description.is_empty() {
@@ -183,7 +257,7 @@ pub(crate) fn cmd_split(
     // select.
     let second_commit = {
         let target_tree = target.commit.tree()?;
-        let new_tree = if args.parallel {
+        let new_tree = if parallel {
             // Merge the original commit tree with its parent using the tree
             // containing the user selected changes as the base for the merge.
             // This results in a tree with the changes the user didn't select.
@@ -199,10 +273,13 @@ pub(crate) fn cmd_split(
         let mut commit_builder = tx.repo_mut().rewrite_commit(&target.commit).detach();
         commit_builder
             .set_parents(parents)
-            .set_tree_id(new_tree.id())
-            // Generate a new change id so that the commit being split doesn't
-            // become divergent.
-            .generate_new_change_id();
+            .set_tree_id(new_tree.id());
+        if !use_move_flags {
+            commit_builder
+                // Generate a new change id so that the commit being split doesn't
+                // become divergent.
+                .generate_new_change_id();
+        }
         let description = if target.commit.description().is_empty() {
             // If there was no description before, don't ask for one for the
             // second commit.
@@ -226,6 +303,106 @@ pub(crate) fn cmd_split(
         commit_builder.write(tx.repo_mut())?
     };
 
+    let (first_commit, second_commit, num_rebased) = if use_move_flags {
+        move_first_commit(
+            &mut tx,
+            &target,
+            first_commit,
+            second_commit,
+            new_parent_ids,
+            new_child_ids,
+        )?
+    } else {
+        rewrite_descendants(&mut tx, &target, first_commit, second_commit, parallel)?
+    };
+    if let Some(mut formatter) = ui.status_formatter() {
+        if num_rebased > 0 {
+            writeln!(formatter, "Rebased {num_rebased} descendant commits")?;
+        }
+        write!(formatter, "First part: ")?;
+        tx.write_commit_summary(formatter.as_mut(), &first_commit)?;
+        write!(formatter, "\nSecond part: ")?;
+        tx.write_commit_summary(formatter.as_mut(), &second_commit)?;
+        writeln!(formatter)?;
+    }
+    tx.finish(ui, format!("split commit {}", target.commit.id().hex()))?;
+    Ok(())
+}
+
+fn move_first_commit(
+    tx: &mut WorkspaceCommandTransaction,
+    target: &CommitWithSelection,
+    mut first_commit: Commit,
+    mut second_commit: Commit,
+    new_parent_ids: Vec<CommitId>,
+    new_child_ids: Vec<CommitId>,
+) -> Result<(Commit, Commit, usize), CommandError> {
+    let mut rewritten_commits: HashMap<CommitId, CommitId> = HashMap::new();
+    rewritten_commits.insert(target.commit.id().clone(), second_commit.id().clone());
+    tx.repo_mut()
+        .transform_descendants(vec![target.commit.id().clone()], |rewriter| {
+            let old_commit_id = rewriter.old_commit().id().clone();
+            let new_commit = rewriter.rebase()?.write()?;
+            rewritten_commits.insert(old_commit_id, new_commit.id().clone());
+            Ok(())
+        })?;
+
+    let new_parent_ids: Vec<_> = new_parent_ids
+        .iter()
+        .map(|commit_id| rewritten_commits.get(commit_id).unwrap_or(commit_id))
+        .cloned()
+        .collect();
+    let new_child_ids: Vec<_> = new_child_ids
+        .iter()
+        .map(|commit_id| rewritten_commits.get(commit_id).unwrap_or(commit_id))
+        .cloned()
+        .collect();
+    let stats = move_commits(
+        tx.repo_mut(),
+        &MoveCommitsLocation {
+            new_parent_ids,
+            new_child_ids,
+            target: MoveCommitsTarget::Commits(vec![first_commit.id().clone()]),
+        },
+        &RebaseOptions {
+            empty: EmptyBehaviour::Keep,
+            rewrite_refs: RewriteRefsOptions {
+                delete_abandoned_bookmarks: false,
+            },
+            simplify_ancestor_merge: false,
+        },
+    )?;
+
+    // 1 for the transformation of the original commit to the second commit
+    // that was inserted in rewritten_commits
+    let mut num_new_rebased = 1;
+    if let Some(RebasedCommit::Rewritten(commit)) = stats.rebased_commits.get(first_commit.id()) {
+        first_commit = commit.clone();
+        num_new_rebased += 1;
+    }
+    if let Some(RebasedCommit::Rewritten(commit)) = stats.rebased_commits.get(second_commit.id()) {
+        second_commit = commit.clone();
+    }
+
+    let num_rebased = rewritten_commits.len() + stats.rebased_commits.len()
+        // don't count the commit generated by the split in the rebased commits
+        - num_new_rebased
+        // only count once a commit that may have been rewritten twice in the process
+        - rewritten_commits
+            .iter()
+            .filter(|(_, rewritten)| stats.rebased_commits.contains_key(rewritten))
+            .count();
+
+    Ok((first_commit, second_commit, num_rebased))
+}
+
+fn rewrite_descendants(
+    tx: &mut WorkspaceCommandTransaction,
+    target: &CommitWithSelection,
+    first_commit: Commit,
+    second_commit: Commit,
+    parallel: bool,
+) -> Result<(Commit, Commit, usize), CommandError> {
     let legacy_bookmark_behavior = tx.settings().get_bool("split.legacy-bookmark-behavior")?;
     if legacy_bookmark_behavior {
         // Mark the commit being split as rewritten to the second commit. This
@@ -258,18 +435,7 @@ pub(crate) fn cmd_split(
         }
     }
 
-    if let Some(mut formatter) = ui.status_formatter() {
-        if num_rebased > 0 {
-            writeln!(formatter, "Rebased {num_rebased} descendant commits")?;
-        }
-        write!(formatter, "First part: ")?;
-        tx.write_commit_summary(formatter.as_mut(), &first_commit)?;
-        write!(formatter, "\nSecond part: ")?;
-        tx.write_commit_summary(formatter.as_mut(), &second_commit)?;
-        writeln!(formatter)?;
-    }
-    tx.finish(ui, format!("split commit {}", target.commit.id().hex()))?;
-    Ok(())
+    Ok((first_commit, second_commit, num_rebased))
 }
 
 /// Prompts the user to select the content they want in the first commit and